+++ continuousintegration/orchestration/create_instance.sh
@@ -43,6 +43,9 @@ keepobjectstore=1
 # By default run the standard test no oracle dbunittests
 runoracleunittests=0
 
+# By default doesn't prepare the images with the previous schema version
+updatedatabasetest=0
+
 usage() { cat <<EOF 1>&2
 Usage: $0 -n <namespace> [-o <objectstore_configmap>] [-d <database_configmap>] \
       [-e <eos_configmap>] [-a <additional_k8_resources>]\
@@ -59,13 +62,14 @@ Options:
   -O	wipe objectstore content during initialization phase (objectstore content is kept by default)
   -a    additional kubernetes resources added to the kubernetes namespace
   -U    Run database unit test only
+  -u    Prepare the pods to run the liquibase test
 EOF
 exit 1
 }
 
 die() { echo "$@" 1>&2 ; exit 1; }
 
-while getopts "n:o:d:e:a:p:b:B:E:SDOUm:" o; do
+while getopts "n:o:d:e:a:p:b:B:E:SDOUum" o; do
     case "${o}" in
         o)
             config_objectstore=${OPTARG}
@@ -114,6 +118,9 @@ while getopts "n:o:d:e:a:p:b:B:E:SDOUm:" o; do
         U)
             runoracleunittests=1
             ;;
+        u)
+            updatedatabasetest=1
+            ;;
         *)
             usage
             ;;
@@ -131,6 +138,19 @@ fi
 
 # everyone needs poddir temporary directory to generate pod yamls
 poddir=$(mktemp -d)
+
+# Get Catalogue Schema version
+MAJOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MAJOR ../../cmake/CTAVersions.cmake | sed 's/[^0-9]*//g')
+MINOR=$(grep CTA_CATALOGUE_SCHEMA_VERSION_MINOR ../../cmake/CTAVersions.cmake | sed 's/[^0-9]*//g')
+SCHEMA_VERSION="$MAJOR.$MINOR"
+
+# It sets as schema version the previous to the current one to create a database with that schema version
+if [ "$updatedatabasetest" == "1" ] ; then
+    MIGRATION_FILE=$(find ../../catalogue/ -name "*to${SCHEMA_VERSION}.sql")
+    PREVIOUS_SCHEMA_VERSION=$(echo $MIGRATION_FILE | grep -o -E '[0-9]+\.[0-9]' | head -1)
+    NEW_SCHEMA_VERSION=$SCHEMA_VERSION
+    SCHEMA_VERSION=$PREVIOUS_SCHEMA_VERSION
+fi
 
 if [ ! -z "${buildtree}" ]; then
     # We need to know the subdir as well
@@ -197,9 +217,9 @@ else
 fi
 
 if [ $keepobjectstore == 1 ] ; then
-    echo "objecstore content (if used) will be kept"
+    echo "objecstore content will be kept"
 else
-    echo "objectstore content (if used) will be wiped"
+    echo "objectstore content will be wiped"
 fi
 
 
@@ -261,7 +281,7 @@ done
 
 echo "Creating pods in instance"
 
-kubectl	create -f ${poddir}/pod-init.yaml --namespace=${instance}
+sed "s/SCHEMA_VERSION_VALUE/${SCHEMA_VERSION}/g" ${poddir}/pod-init.yaml | kubectl create --namespace=${instance} -f -
 
 echo -n "Waiting for init"
 for ((i=0; i<400; i++)); do
@@ -290,6 +310,7 @@ if [ $runoracleunittests == 1 ] ; then
     kubectl get pod oracleunittests -a --namespace=${instance} | egrep -q 'Completed|Error' && break
     sleep 1
   done
+  echo "\n"
 
   kubectl --namespace=${instance} logs oracleunittests
 
@@ -303,7 +324,6 @@ if [ $runoracleunittests == 1 ] ; then
   # database unit-tests were successful => exit now with success
   exit 0
 fi
-
 
 echo "Launching pods"
 
@@ -425,16 +445,20 @@ kubectl --namespace=${instance} exec ctaeos cat /etc/eos.keytab | kubectl --name
 kubectl --namespace=${instance} exec ctaeos cat /etc/eos.keytab | kubectl --namespace=${instance} exec -i client --  bash -c "cat > /etc/eos.keytab; chmod 600 /etc/eos.keytab"
 echo OK
 
-
-echo -n "Waiting for cta-frontend to be Ready"
-for ((i=0; i<300; i++)); do
-  echo -n "."
-  kubectl --namespace=${instance} exec ctafrontend -- bash -c 'test -f /var/log/cta/cta-frontend.log && grep -q "cta-frontend started" /var/log/cta/cta-frontend.log' && break
-  sleep 1
-done
-kubectl --namespace=${instance} exec ctafrontend -- bash -c 'grep -q "cta-frontend started" /var/log/cta/cta-frontend.log' || die "TIMED OUT"
-echo OK
-
+# In case of testing to update the database using liquibase.
+# If the previous and new schema has different major version, the ctafrontend will crash,
+# so it's not necesary to check if it's ready
+NEW_MAJOR=$(echo ${SCHEMA_VERSION} | cut -d. -f1)
+if [ "${MAJOR}" == "${NEW_MAJOR}" ] ; then
+  echo -n "Waiting for cta-frontend to be Ready"
+  for ((i=0; i<300; i++)); do
+    echo -n "."
+    kubectl --namespace=${instance} exec ctafrontend -- bash -c 'test -f /var/log/cta/cta-frontend.log && grep -q "cta-frontend started" /var/log/cta/cta-frontend.log' && break
+    sleep 1
+  done
+  kubectl --namespace=${instance} exec ctafrontend -- bash -c 'grep -q "cta-frontend started" /var/log/cta/cta-frontend.log' || die "TIMED OUT"
+  echo OK
+fi
 
 echo "Instance ${instance} successfully created:"
 kubectl get pods -a --namespace=${instance}