+++ continuousintegration/orchestration/run_systemtest.sh
@@ -60,6 +60,7 @@ Options:
   -a    additional kubernetes resources added to the kubernetes namespace
   -U    Run database unit test only
   -C    Cleanup leftover kubernetes namespaces
+  -u    Prepare the pods to run the liquibase test
 
 Create a kubernetes instance and launch the system test script specified.
 Makes sure the created instance is cleaned up at the end and return the status of the system test.
@@ -72,7 +73,7 @@ exit 1
 # always delete DB and OBJECTSTORE for tests
 CREATE_OPTS="-D -O"
 
-while getopts "n:d:s:p:b:e:a:B:t:kDOSUC" o; do
+while getopts "n:d:s:p:b:e:a:B:t:ukDOSUC" o; do
     case "${o}" in
         s)
             systemtest_script=${OPTARG}
@@ -121,6 +122,9 @@ while getopts "n:d:s:p:b:e:a:B:t:kDOSUC" o; do
             ;;
         C)
             cleanup_namespaces=1
+            ;;
+        u)
+            CREATE_OPTS="${CREATE_OPTS} -u"
             ;;
         *)
             usage